+++ core/src/main/java/com/asml/apa/wta/core/model/Resource.java
@@ -1,8 +1,9 @@
 package com.asml.apa.wta.core.model;
 
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Setter;
+import lombok.Getter;
 
 /**
  * Resource class corresponding to WTA format.
@@ -16,8 +17,8 @@ public class Resource implements BaseTraceObject {
 
   private static final long serialVersionUID = 3002249398331752973L;
 
-  @Setter
-  private static String schemaVersion;
+  @Getter(value = AccessLevel.NONE)
+  private final String schemaVersion = this.getSchemaVersion();
 
   private final long id;