+++ core/src/main/java/com/asml/apa/wta/core/model/Workflow.java
@@ -1,9 +1,10 @@
 package com.asml.apa.wta.core.model;
 
 import com.asml.apa.wta.core.model.enums.Domain;
+import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
-import lombok.Setter;
+import lombok.Getter;
 
 /**
  * Workflow class corresponding to WTA format.
@@ -17,8 +18,8 @@ public class Workflow implements BaseTraceObject {
 
   private static final long serialVersionUID = 9065743819019553490L;
 
-  @Setter
-  private static String schemaVersion;
+  @Getter(value = AccessLevel.NONE)
+  private final String schemaVersion = this.getSchemaVersion();
 
   private final long id;